requirements.txt
rich
supabase
python-dotenv
Flask
Flask-Cors